Rename scale to scaleHelpers
