+++ src/utils/scaleHelpers.js
